Remove unused tracking methods for calling a method on the next page
As part of a refactor to reduce the number of public functions we
expose on the `StaticAnalytics` object, we have removed all calls
to the `callOnNextPage` method. This method was overly-generic, allowing
any arbitrary method to be called on the next pageload, when in fact
we just wanted to set a custom dimension for the next tracked pageview.
Custom dimensions can instead be set by using the
`setOptionsForNextPageView` method.