Remove unused tracking methods for setting custom dimensions
As part of a refactor to reduce the number of public functions we
expose on the `StaticAnalytics` object, we have removed all the old
public methods for setting custom dimensions at a session level.

Instead, custom dimensions should be passed directly into the tracking
methods as part of the [fields object](https://developers.google.com/analytics/devguides/collection/analyticsjs/field-reference).

Note that the public `setDimension` method is being left in place
because it is still being used by a large number of applications across
GDS.