Make steps not images and accessible via css
So the Initial aim was just to add some css so i could access the numbers through css
i then dicovered a few issues with them. The first was the excessive amount
of css it generated due to having two images. The next one was that the
content was in an image, so didnt scale properly or accessible in the dom.

due to the concerns of adding extra html in, i decided to use before and after
which seems to be reasonably well supported. It's an improvement over what
was there, but maybe not perfect. it seems to be supported by a lot of assistive
technology, but not maybe it all, which is an improvement over none.