Improve mechanism for selecting which survey to run
The previous algorithm was to choose the default survey and then ask every
other survey defined in `smallSurveys` in turn if it was active. If the
survey says it is then we record it as the active survey and move on to
the next survey to ask if it is active. Once we have a survey then we
check if the survey has already been completed, and roll the dice to check
against the frequency.  Only then do we show it.

There are a couple of problems with this algorithm.  First; if we have two
or more active surveys that could run on a given page we always pick the
one defined last in smallSurveys.  This means that on that page we'll
never show the ones defined earlier.  Specifically this means that we'll
never show the default survey on a page that another survey is running on.
This is not about showing multiple surveys on a page at one time, just that
across all page views we will only ever show the last defined one.  Second;
by choosing the active survey before we check if it can be shown we run the
risk of picking a survey that we can't actually show and thus showing no
survey at all when another survey we could have picked would be showable.

To solve this we select all possible small surveys for the page looking at
their start and end times, and at the `activeWhen` function.  We add the
default survey to this list and then filter out any that can't be shown
because they have already been completed, or their frequency doesn't
match.  If after doing this we have more than one survey we roll the dice
to pick one of them.  It should be safe enough to do this as we have no
other way of choosing which should run, and we've already checked their
frequencies.  This means that a high frequency survey that runs on a page
won't always be shown in favour of a low frequency survey that also runs
on that page, and also that the last defined survey won't always be the
winner either.