Enable Google Analytics Enhanced Ecommerce for search
Track search results using Enhanced Ecommerce, which provides analytics
on views (impressions) and clicks of lists of items.

We are only sending the `id` for each view/click, with additional metadata
being uploaded as GA product data. We have done this to ensure we don't bloat
the data request for the pageview event which has a max size of 8k.