Use `.and.returnValues` correctly in js specs
To return different values for multiple calls to a function we need to use
`.and.returnValues` not `.and.returnValue`.  This was causing specs to
pass for the wrong reasons.  Happily using the correct API doesn't cause
any breakages, but it might if they are used for inspiration of future
tests.