Add limit for how many times a survey can be seen
We increment a counter each time a survey is shown to the user and store
this in a session cookie.  When picking which survey to show to the user
is is discarded if it has been shown to the user more than 2 times within
their session.  The counter is incremented separately for each survey;
showing survey X a couple of times does not mean they don't get to see
survey Y.

This allows us to ramp up the frequency that a survey will be shown
without annoying frequent visitors who have ignored the survey multiple
times.