Provide a cooloff in days for survey seen too many times cookies
By default we set the survey seen too many times cookie as a session
cookie that will last until the browser is closed.  However, if
`seenTooManyTimesCooloff` is specified in the survey definition we use
that value as the number of days to store the cookie for instead.  This
allows us to tweak the behaviour of the functionality for hiding a survey
from a user that has already seen it.