Make the default survey appear 1 in 6 times
Now that we have functionality to stop showing a survey over and over to
a user who has not interacted with it we can be bolder about how often
we show the default survey.  This should mean we get more results and
better data about what people think of the platform.