Replace `activeWhen` function with static arguments in surveys
We extract the functionality that has been used in most of the `activeWhen`
functions defined on surveys to date, and make them part of the main survey
object. The survey definitions now describe the parameters for these functions
instead of defining the functions themselves.

All previous surveys have used OR to combine their rules for when
to show the survey, so that's what we do here.  We've also allowed
for inverting the functionality from "show if any of these match" to
"don't show if any of these match" via a `matchType` param that can
be set to `exclude`. Only one survey has run on an exclusion basis
so far, so we don't know if we'll ever need something that has rules
like "show on this section, but not this page in that section". We
can build that in later if we need it.

Moving from a function to params does limit our capability for how to
decide to show each survey, but the upside is that we have tests for all
our matchers now.