Add a new survey for HMRC user research
A 1 in 20 survey that runs across a bunch of HMRC pages.  Each page needs
to have a different `utm_campaign` value for the survey url, and it's not
easily generated from the path, so we add a simple map function to
translate between the path and the `utm_campaign` value.  This runs at
load time rather than survey invocation time because we don't (currently)
support functions for url values.  This does mean that the survey url will
not contain a `utm_campaign` value if the path is not one of the ones
matched by the list of urls.  This isn't a problem though as we use the
same mapping function to test if we should show the survey in the
`activeWhen` function of the survey, meaning that no path that results in
an empty `utm_campaign` value will ever show the survey.

See: https://trello.com/c/t8HJ6lFM/190-govuk-survey-user-research-panel-invite-hmrc