Fixing misaligned warning sign on mobile
Fixes #1061 . Previously, the warning sign on transaction pages was misaligned on mobile due to the default margin being overridden by a separate style. It looks like this separate style was being wrongly applied to multiple elements when it should have just been applied to the first paragraph of any transaction page, to reduce the distance between the first heading and p.

This fix makes the CSS selector more specific to make sure this selector only applies as described above, and no longer effects warning signs.