Track total number of topic sections
We are adding tracking events to old navigation pages in order to track
how people use them. We have added tracking events when users click on
topic links, but we also need information on how many sections there are
on those pages, so that the custom dimension is reported accurately to
Google Analytics.

This commit makes sure we set the custom dimension 26 with the correct
number of topic sections.

This is related with the GitHub PR [1] and the Trello ticker [2].

- [1] https://github.com/alphagov/collections/pull/361
- [2]
  https://trello.com/c/5YQVnGVr/32-add-total-links-total-sections-counting-to-whitehall-navigation