Track document collection links and sections
This commit makes sure we track how many links there are on a document
collection. It also makes sure we track the number of sections
accurately, by first checking if we are in a document collection page.
Since document collections are content items and therefore might have
related links, the order in which we check for these counts matter.