Add `link_text` param to EmergencyBanner::Deploy
Emergency banners have the option of displaying a hyperlink for more
information. The hyperlink text is currently hard coded to "More
information" in the view. We are changing this so that the text can be
user defined.

This commit sets the groundwork by writing an additional `link_text`
parameter to Redis to hold the text we wish to display for the
hyperlink.