Only send GA events in webchat for state changes not every message
This webchat implementation polls the iframe every second to see if it is
ready to let people talk to an advisor.  If it sees that it is then we
send an "offered" event to GA.  It'll do this once a second until the user
accepts, or rejects, or the webchat system changes state to indicate that
no-one is available.  This flood of messages is clogging up our GA hits so
we want to reduce the events sent.

We take inspiration from the government-frontend implementation of webchat
and only send the GA event when we detect a change in the state.  We only
record the last state so if it flip flopped between available and closed
every second we'd still send an event every second, but if it remained
open for 5 minutes before switching over we'd only send 2 GA events.

To help with tracking the state changes we introduce tracking of the
"closed" event (with status "unavailable" like the government-frontend
version).  Otherwise we'd only track a state change from "offered" to
"error" and not from "offered" to closed to "offered" again.