Remove leading period from custom survey postscripts
The custom call to action already ends with a period so we don't need
to add one to the postscript.  Even if we wanted to have the period out
side the all to action text, there's a big gap between it and the
postscript text when they're actually rendered rendered so we shouldn't
do it this way.