Reconfigure smallSurveys to use arguments not functions
In [#1078](https://github.com/alphagov/static/pull/1078) we changed the
behaviour of the surveys system to make the `activeWhen` property a list
of arguments instead of a function.  However, we forgot to reconfigure
the existing small surveys to use this functionality.  This results in
all the surveys behing potential candidates for every page.