Send state and language to GA
Sends the publishing state of the document and the language it's written in to Google Analytics. It will do this on every pageview event for publications, if it finds the state and/or the current language.

The purpose is to gather more data about which documents are viewed the most when filtered by publishing state and language.