Add `link_text` property to EmergencyBanner::Display
Expose the link_text property in EmergencyBanner::Display so that we can
subsequently use it in the front end.

If no link is present then we don't need link_text and set it to nil.

If there is a link, then we should display either the link_text (if it
exists) or a default of "More information".