Refactor how we calculate section and link counts
This commit moves the logic of determining how many sections and links
are on a given page to a new module, which has more granular functions
to extract all the necessary bits from a page to make a decision.

This refactoring aims at making it clearer what the steps are. Ideally,
each rendering application would set these counts in a meta tag and this
code would only read from it, but it's a bit more complicated than that,
especially for JS-driven finders.