Add SRI attributes to CSS and JS tags in templates
This lets the browser check that nothing has been tampered with between
us and them.  Note that a bunch of the assets, and html templates for
static actually come from govuk_template, and we'll bump the version of
that in another commit to get SRI applied to those assets too.