Rename "current" to "not withdrawn"
For: https://trello.com/c/KXyHx427/212-rename-current-to-not-withdrawn-when-sending-data-to-ga-on-publication-status
This is aimed at clearing up some confusion in GA data where documents that are marked as current/withdrawn may actually be in a third state where they are not relevant anymore but have not yet been withdrawn.
For this purpose, we're renaming the word "current" to "not withdrawn" to keep the current binary classification and reduce confusion.