Send the client ID inside a custom dimension
For: https://trello.com/c/n22tcj9Q/206-custom-dimension-google-analytics-client-id-and-session-id

In order to track user journeys in GA, it's been requested that we send the client ID as a custom dimension.
The recommended way to extract the client id can be found here: https://developers.google.com/analytics/devguides/collection/analyticsjs/cookies-user-id?hl=en#getting_the_client_id_from_the_cookie

Implementing the recommended way implies expecting a tracker, which may or may not have been fetched at that point. Because our method is asynchronous, it might be called before trackPageView and not have a tracker object to query for the client ID.
In order to avoid this, we've put the call for trackPageView inside the asynchronous method for setting the client ID. This will guarantee that these two methods are called in the right order.
We've also had to move the rest of the plugin loading after the trackPageView event in order to ensure the same order will be kept (although ultimately trackPageView might still be called asynchronously).