Send client ID with survey currentPath param
For: https://trello.com/c/F5wcNrsV/219-clientid-pushed-into-survey-urls-as-custom-data

In order to track the user journey when a user completes a survey, we can use the client ID to identify a session for a user and check all the pages he had visited before arriving to our survey. This will provide more detailed data about a user's browsing habits and will help to separate the data out, since the info collected on surveys is anonymous.