Add new survey for education and childcare content
Much of this content is survey-limited because it has the beta navigation
banner which blocks surveys from appearing.  However, not all of it does
and the team are keen to get this survey out regardless.  If it turns out
that not enough responses are generated we might look at removing the beta
banner block on surveys, but that needs more discussion to avoid survey
fatigue on those pages.