Avoid talking to slimmer when doing translations
Requests to GovukPublishingComponents will add the slimmer backend to our
i18n backend chain.  This backend will try to make a network request to
fetch locales, and do the translation.  We don't need to do that here
because ultimately the request would come back to this app and look at the
same i18n files that the previous backend in the chain has looked at.  We
can't remove the slimmer i18n backend from the chain because it is only
added after a request is made to the govuk publishing components
controllers, and it's added again if it's not present.  The solution we
use is to replace the slimmer i18n backend with a null backend that will
do nothing.

This came to our attention because the metadata_test would occasionally
fail when running tests.  This test case includes a locale test, and
would pass if it was run before the component_guide_test but fail if it
was run after them because the component_guide_test makes a request to
the govuk publishing components and that sets up the slimmer i18n backend.