Change the survey urls for the hmrc jul 2017 surveys
The links we were originally given have stopped working for some reason
so we need to change them all.  The original links all had the same
query string except for the value of the `utm_campaign` param, but these
new ones have different `id` and `utm_medium` values so our mapping
function now becomes a bit more cumbersome.  We may want to think about
talking to hmrc in the future to find out why it's important that all
the query string stuff is so different, isn't it enough to vary only
by the utm_campaign (or some other variable) as this'll make it easier
for us to have one survey that contains the definition for all their
links.