Let surveys target pages via govuk:themes as well as govuk:section
In the new navigation the govuk:themes meta tag will contain the "root" of
the hierarchy, similar to how govuk:section does (when present).  If a
survey has a `section` attribute in the `activeWhen` we'll check it
against both the govuk:section and govuk:themes value.  There is a gotcha
here because `govuk:section` will be "Human Readable" but `govuk:themes`
will be `machine-readable`, but as long as we only check individual words
we should be ok.