Remove old HMRC webchat
For: https://trello.com/c/hWTtb0OZ/253-remove-old-instance-of-webchat

This implementation of webchat uses the old HMRC webchat so we want to get
rid of it.  There is a newer implementation included in
government-frontend that we could replace this one with, but we realised
that most of the pages the older webchat version is active on link to an
HMRC contact page.  All the HMRC contact pages have the new webchat
implementation on it so there's no need to have the old webchat.  For those
that don't link to an HMRC contact page we decided it wasn't a problem to
not have webchat.