Add new survey for hmrc covering import/export users
This is a 1 in 6 frequency that lasts for 3 months across a handful of
pages relating to import/export.  We have to use a function again to set
the url for the survey because for the 9 urls we hit we want 3 different
survey urls.  We may need to think about allowing the survey url to become
a function as we've had to do this a few times now, and it'd be better if
the full survey definition was contained in a single object rather than
an object, and a function defined elsewhere.  The alternative is defining
seaprate surveys for each url, but that seems worse somehow.