Remove stray slash in hrmc manuals survey
This survey should work on all paths under hmrc-internal-manuals, but we
left a slash in our regexp matcher that means it would only match entries
with a double slash (e.g. /hrmc-internal-manuals//foo instead of
/hrmrc-internal-manuals/foo).

We probably need to think about an abstraction for defining these matches
to make it hard to make this kind of mistake in the future.