Set the language custom dimension in customDimensionsFromDom
It used to be done in customDimensionsFromMetaTags, but the value is not
coming from a meta tag so that's not really where it belongs.  We also do
a similar reorganisation of the tests and create explicit tests for
setting the custom dimension when the dom element is or isn't present.