Make meta tag custom dimension specs more explicit
We already had explicit tests for the custom dimensions that had default
values, so we add explicit tests for the ones that don't have default
values too.  We also add a couple of missing default value tests.  This
allows us to remove one of the tests that relies on the length of the keys
in the `pageViewObject` which is fairly brittle.