Clarify what getPageViewObject() is doing
There's a bunch of magic numbers and assumptions in this method about
how many ga() functions have been called, in what order, and for what
reason.  This is fine, but it's hard to work out so we leave some more
explicit comments, and introduce an expectation that will fail if our
assumptions are false.