Allow configuring stripping postcode PII in analytics via a meta-tag
Stripping PII is a function provided by the GOVUK.Analytics module from
govuk_frontend_toolkit in version 7.3.0 which we upgrade to in this
commit.

By default it strips emails from all arguments passed to GA, but it can
also strip Postcodes if you tell it to.  It doesn't do postcodes by
default because they're much more likely to cause false positives.
There's no easy way for apps using static to run some code before
analytics/init.js is run to let them pass arguments or configure the
object before the initial trackPageVie is called, so we look for the
presence of a meta-tag called "govuk:static-analytics:strip-postcodes".
If it's there, we will tell GOVUK.Analytics to strip postcodes, if it's
not we will tell it not to.  It doesn't matter what the content of the
meta-tag is, just if it's there or not.