Always add the strip-postcodes meta tag for some content items
Smart Answers and search pages are the formats where we are likely to have
postcodes in the GA data.  Smart Answers because some of them have a
postcode entry question that forms part of the URL, and search because we
see people searching for their postcodes fairly regularly.

Unfortunately this approach is hampered by the fact that smartanswers
report the wrong format at the moment. Smart answers are split into two
documents; the start-page at `/slug` which is a `transaction` and rendered
by frontend, and the interactive Q&A part at `/slug/y` which is a
`smart_answer` and rendred by smart-answers. Unfortuntately smart-answers
loads the `/slug` content item, not the `/slug/y` content item so it
reports the format as transaction, not smart_answer. Until this is fixed
we want to be able to explicitly say in smartanswers that we should strip
postcodes, so we allow a `strip_postcode_pii` option to be provided to the
component that will always add the meta tag, regardless of the
document_type of the content_item. This may come in useful for other one
off pages where we need to strip postcodes.