Add cross domain analytics tracking via data attributes
This allows a generic method of tracking links on GOV.UK.

This will be useful for when our govspeak extension is used in content,
where only content designers can author what goes into the body of the page.

Here we're using native DOM methods rather than jQuery fragments because, for
whatever reason jQuery produces a fragment that does not allow inspection of hostname.

So rather than add additional overhead to the users' javascript, the tests are a little more verbose.