Enables custom dimension for uploads
For: https://trello.com/c/SYgBUl4q/258-develop-an-upload-to-ga-of-link-checker-data

This is a follow up to a discussion which arose from the fact that `govuk_frontend_toolkit` is using this custom dimension to duplicate a url into a custom dimension.
However, because `govuk_frontend_toolkit` is used in other places besides `gov.uk`, it would inadvertently use their custom dimension as well.

So we've made this behavior optional and we're sending the custom dimension via `static`. If the custom dimension isn't present, the behavior will be turned off.
If it's present, it will use the custom dimension given to duplicate the url information into that custom dimension.

This will later be used to join the url data in the CD with information from a Google Analytics custom upload. Unfortunately, we can only join up with the custom upload info via custom dimension.
To achieve this join, we've had to duplicate the url information that is normally kept in the `event action` into a custom dimension.