Re-add old taxonomy styles for better deployment support
Static template code can sit in the application caches long enough
for the wrong combination of CSS and template code to render on page.
This leads to a broken looking component. So add the old styles to support
the deployment of the recent update to the taxonomy sidebar.
These styles can be removed once the applications are up to date.