Adds CSS for mobile version of survey
This will add special styling for surveys displayed on tablets or mobile screens.

`@include media-down(mobile)` is used to specify both tablet (<=800px) and mobile(<=640px), as opposed to using `@include media(mobile)` (<=640px).

For surveys displayed on mobile (<= 800px), this increases the size of the “Close” button and adds some spacing to the survey title. It also makes some minor font changes.