Adds TLS cookie check and new TLS survey
This survey will be displayed on both mobile and desktop, for people using a TLS version < 1.2 (considered insecure).

We fetch the version from the `TLSversion` cookie, where it should be present as `TLSv1.x` | `unknown`. In some case, the cookie can be missing.
We don't want to show the survey to people who don't have the cookie set or have the value stored as `unknown`.

We’ve also added an optional `allowedOnMobile` param on the survey.
When `allowedOnMobile` is:
- set to `true`, the survey will be displayed on both desktop and mobile
- set to `false` or omitted altogether, the survey will only be displayed on desktop.