Add CTTUK tree test survey
The original requirements for this survey was that it should appear on
/browse/visa-immigration and 2641 other unique paths. However, it was
unfeasible to include this list of paths in the survey.js file, which
would considerably increase the size of the file (from ~32kB to 296kB).

As an alternative approach, we have chosen a subset of organisations and
a handful of paths that receive a high volume of pageviews but would not
be covered by the organisations, for this survey to active on.