Add performance considerations to surveys documentation
When talking about the survey in #1283 the original version contained ~3000
path comparisons.  We realised that this could be problematic for browsers
because we'd have to run all those regex comparisons on every request while
the survey was running.  We haven't done any benchmarking so it's not clear
when we need to worry about performance and what we can get away with, but
in general it makes sense to keep the number of comparisons low on a survey
if possible so we've added some notes about this to the documentation.