Mark all custom dimensions as PII safe
All of these come from data that we control and have put on the page
explicitly.  None of them come from data entered by the user.  This
means it's safe for us to mark them as PII Safe and avoids any false
positives for emails or postcodes being stripped out of data we want
to track.  For example, it's fairly likely that a govuk content id
could contain something that looks like a postcode and if we strip
it out, we break the content id.

The PIISafe object comes from govuk_frontend_toolkit v 7.4.0 so we
upgrade to that in this commit too.