Strip PII from data sent to GA by ecommerce.js
This code deosn't use the `GOVUK.StaticAnalytics` or `GOVUK.Analytics`
abstractions when talking to GA so it didn't benefit from the PII
stripping code in there. It tracks events which may include user
provided data, and it sends the raw search query which also may
include user provided data so we need to strip PII from those.  For
events we change ecommerce over to use the `GOVUK.StaticAnalytics`
abstraction for tracking events.  For the search query which is sent
to a GA method we don't have an abstraction for, we have to explicitly
strip the PII ourselves, this means we need to expose the
`GOVUK.Analytics.stripPII` method on `GOVUK.StaticAnalytics`.