Fix analytics for empty pages part of step navs
We don't want to add analytics where there are no related step navs.

At the moment, content with no part_of_step_navs links will be defaulted to an
empty array which results in an empty string ("") from the map. The empty
string is truthy which is why the tests are failing.