Only add linked trackers once
If we attempt to use cross domain tracking on more than one button
on a single page the linked trackers are added twice. This means
that the underlying ga library throws a javascript error.
This commit appends the tracker names to an array to check for
repeat calls to the same tracker.

It also changes the data attribute necessary for tracking an additional
event.