Memoize linked trackers in the Modules namespace
GOVUK.Modules initializes and starts a module for every element
with a data-module attribute. This causes errors for cross-domain-tracking
if two elements should send data to the same tracker because the
underlying ga library doesn't like duplicated config calls.
The trackers are memoized by name at the Modules level so that new
instances of CrossDomainTracking can be initialized and started
without reconfiguring the same ga tracker.