Deduplicate trackers regardless of context
The cross domain tracking module can be applied to a block level element
containing matching elements or to an element suchas an anchor tag.
In the case of multiple start buttons GOVUK.Modules attempts to initialize
trackers on a per button basis, so it's important that both recursive and
explicit element contexts are covered.
This means moving the check for duplicate trackers into the addLinkedTrackerDomain
function as this is called for either element context lookup.