Specify which JavaScript assets to build
This changes the wildcard entry which meant all js files were being
built as assets. This was a crazy amount of files for how many are
needed (5) as it was attempting to build every single file in the
node_modules directory from govuk_publishing_components.

There is a risk here that the removal of all the files that are
published either break static or some other app. Static should only
break if I've missed a file that's needed, if other apps are referencing
the files here then that's a real pain - but that seems unlikely due to
asset pipeline fingerprinting.

This is intended to fix the issue that this project can't build with
versions of govuk_publishing_components past 11 as it chokes on files in
ES6 syntax.

This reduces the number of files generated as assets by static from 931
to 383.