Make barchart govspeak enhancement resilient to re-initialisation
This barchart initialisation is vulnerable to breaking a component if
the initialisation is called twice. As the code for this is moving to
govuk_publishing_components there is going to be a period of time where
this can quite easily occur while apps are being updated and deployed.

This updates this initialisation code to mark when a chart has been
initialised and be able to ignore being run on charts that have already
been initialised.