Reverting change in search box action
We want to retain the behaviour where empty searches go to the search page
at `/search`.  By changing the action to go to `/search/all` we stopped this.

The `/search` page redirects searches with keywords to go to `/search/all`
and adds the `order=relevance` param.  We added the hidden input here to
replicate that behaviour, but if we're reverting then we no longer need it.

I've added a comment to remind future me to not assume that it's trivial.