Add direct accessibility link to footer
##What
Add a direct link to the accessibility page from the GOV.UK footer _(the list Help | Cookies | Contact etc)_  (you currently have to navigate through the ‘Help’ section). 

It should be added between "Contact" and "Terms and conditions". 

##Why
This makes our accessibility link more prominent . Monitoring body advice is that the link should never be more than one click away from the user.

##Background

EU Accessibility Directive : _“A link to the accessibility statement should be prominently placed on the homepage of the website or made available on every web page, for example in a static header or footer.”_

This applies to new websites from September 2019. Strictly speaking, it doesn’t apply to existing websites until September 2020, but we’ve committed to making GOV.UK bit of an exemplar - so would be good to do it ahead of the September 2019 deadline.