Ability to track click events with a custom label
At the moment all instances of tracking click events in GA
are sent with an eventLabel of 'Results',
eventhough we have the ability to track different lists.

This work extends that by providing the option
to specify the label sent to GA when clicking a link in a specific list.

This can be done by adding an additiional  'data-track-click-label'
attribute to the element where we initialise list tracking.