Add versioning mechanism to global banner
The global banner currently disappears after it has been viewed 3 times or if a user manually dismisses it.

This banner could be used in the future for other events, and we therefore may want to "reset" the banner so users who have previously seen it or manually dismissed it will be able to view it again.