Handle global_bar_seen cookie not being set
Some of our Smokey tests don't seem to include the global bar, and so the JS included in the global bar view is not executed, including setting a cookie if it doesn't yet exist.

This commit replicates that logic in the global-bar.js file too, so if the cookie is not set, it will set a basic one. The next time the user navigates, that cookie will be updated with the correct version etc.