Track ecommerce variant if set
We use ecommerce tracking to know what people click on in search, and
where in the results list that is: when sorting by relevancy, we want
higher up results to match what the users expect.  But relevancy isn't
the only way to sort search results, in which case we wouldn't
necessarily want or expect clicks to be clustered at the top.

So we can see how sort order affects clicks, and to avoid lumping
different use-cases together, we're going to use the selected sort
option as a product variant.