Add custom meta tag to the dimensionsMap
We'd like to track spelling suggestions on page view. Previously,
we set the value via javascript, but it was too late as the
pageview object payload was already sent.

Adding this custom meta tag to the allowed map, will enable us
to track if suggestion are displayed on page view.