Remove close date logic from banner
We previously added logic in https://github.com/alphagov/static/pull/1934 to automatically take down the banner at a specified close date.
However, this didn't work as expected due to templates being cached when they are requested by Slimmer. Bypassing this caching would be more challenging and as there isn't an immediate use case for this work (we can manually remove the banner instead) we're going to remove this logic for now.

If this feature is requested in the future, we now know that this is more difficult to implement than we first thought and can scope accordingly.