Add ability to make global banner permanent
The default behaviour of the banner is to show for 3 pageviews, and then automatically hide. It can also be manually dismissed by the user.

We want the ability to have the banner be permanent, i.e: it will always show, regardless of the number of pageviews. This won't affect the user's ability to manually dismiss the banner.