Refactor global bar Javascript
The view contained minified Javascript, and the tests for this logic required devs to copy and paste the minified JS into the test file.
It wasn't clear that this needed to be done, so a developer making changes to the view would assume either 1) the file was untested or 2) their change hadn't broken anything, whereas in reality the test was running against old logic.

This commit moves the JS into a separate JS file so it can be directly tested.